refactor(rdf): use `Optional` in `DeserializationReport`
