+++ src/main/java/ch/psi/ord/core/RoCrateImporter.java
@@ -5,8 +5,8 @@ import ch.psi.ord.model.NoEntityFound;
 import ch.psi.ord.model.Publication;
 import ch.psi.ord.model.ValidationReport;
 import ch.psi.ord.model.ValidationReport.Entity;
+import ch.psi.rdf.DeserializationReport;
 import ch.psi.rdf.RdfDeserializer;
-import ch.psi.rdf.RdfDeserializer.DeserializationReport;
 import ch.psi.rdf.RdfUtils;
 import ch.psi.scicat.client.ScicatClient;
 import ch.psi.scicat.model.v3.CountResponse;
