+++ src/test/java/ch/psi/rdf/RdfSerdeTest.java
@@ -6,7 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import ch.psi.rdf.RdfDeserializer.DeserializationReport;
 import ch.psi.rdf.TestClasses.PrimitiveTypes;
 import io.quarkus.test.junit.QuarkusTest;
 import java.time.Instant;
